Move DiffHunk into DiffFile
